Adding Runner package
The purpose of the runner package is to monitor the amount of time
a program is running and terminate the program if it runs too long.
This pattern is useful when developing a program that will be
scheduled to run as a background task process or as a cron job.
It stops the execution ether if the time runs out or if the
system interrupt signal is received.