Adding Work package
The purpose of the work package is to create a pool of goroutines
that will perform and control the amount of work that gets done
concurrently. This is a better approach than using a buffered
channel of some arbitrary static size that acts as a queue of
work and throwing goroutines at it.
Unbuffered channels provide a guarantee that data has been exchanged
between two goroutines. This approach of using an unbuffered channel
allows the user to know when the pool is performing the work,
and the channel pushes back when it can’t accept any more work
because it’s busy. No work is ever lost or stuck in a queue
that has no guarantee it will ever be worked on.