Adding Resource package
The purpose of the resource package is to provide a pool with a set
of resources that can be shared and individually used by any number
of goroutines.
This pattern is useful when you have a static set of resources
to share, such as database connections or memory buffers.
When a goroutine needs one of these resources from the pool,
it can acquire the resource, use it, and then return it to the pool.